upgrade to latest version of golangci-lint
The version we currently have doesn't support Go 1.22 properly, so it
throws a bunch of warnings locally when trying to run it with the
latest Go version installed.  Just running the latest version locally
also doesn't quite work, since it throws a bunch of errors from the
perfsprint linter, which seems to have gotten stricter.

Upgrade to the latest version of golangci-lint, and fix all the errors
we're getting from it.  Mostly done via `perfsprint -fix`.