Allow setting version during generation (#15970)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

Some schemas can't contain the correct version as they need to be
committed. Therefore allow the option to set the version in the schema
at the point of generation.

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`
- [ ] I have added tests that prove my fix is effective or that my
feature works
  - There appears to be no existing tests for this command.
- No changelog required as this is still a hidden command for internal
use.