Apply ignoreChanges to refresh operations
Makes two changes to the handling of ignoreChanges:
1. Stores the ignoreChanges configuration for each resource in the state file
2. Applies the ignoreChanges configuration from the state file as part of processing a refresh

This allows users to use `ignoreChanges` to prevent refresh diffs being reported when changes happen in the cloud provider that they do not want to bring back into their Pulumi state.

This can be used to acknowledge that part or all of a resource may change in the cloud provider, but the user is okay with this and doesn't want to be notified about it during refresh.

Importantly, this means that the diff won't be reported, but also that the changes won't be applied to state.

The ignoreChanges option when applied to a refresh operation applies to *both* input and output properties of the resource.  This is different from the behavior during a normal update operation, where ignoreChanges only applies to input properties.

The special "*" property path ignores all properties, but also ignores the case where the resource is no longer found in the cloud, and instead keeps the resource with all of it's existing input and output properties.

Because the program is not run for refresh operations, users must apply the `ignoreChanges` via a `pulumi up` first, and then future `pulumi refresh` or `pulumi refresh --preview-only` oeprations will ignore the specificd changes.  This is unfortunate, but the only option we have unless/until Pulumi fundamentally changes it's approach to refresh and decides to run the program as part of computing the resource state (and options) to use during the refresh.