[sdk/python] Workaround lazy module regression (#16038)
A regression was introduced in Python 3.11.9 and 3.12.3 that breaks lazy
loaded `config` modules in generated Pulumi provider Python SDKs. It's
unclear if this is going to be addressed in subsequent patch releases of
Python, so we workaround the problem by using a copy of the Python
stdlib's `_LazyModule` and `LazyLoader` classes that don't have the
regression, as suggested by the Python maintainers.

Fixes #15882