remove fsutil.Chdir when running deployment
Remove the Chdir we currently do in the engine.  This probably was
required at one point, but it no longer is.

The engine itself doesn't do anything with the CWD.  At the time the
deployment is running all the data it needs has already been read, and
all it does at this point is communicated with the language plugin and
the providers.

Those providers are already started with the correct CWD passed in to
its execution (see https://github.com/pulumi/pulumi/blob/f1f4d223f4ad275f7ade1e34b6229931cf7636ac/sdk/go/common/resource/plugin/plugin.go#L377).

And as these subprocesses are executed as binaries, they are
unaffected by the Chdir anyway.  Remove the Chdir, as it unlocks
further improvements.