display: only hide replacement steps in diff
When displaying diff events, we currently try to hide non-logical
replacement steps unless we specifically enable showing them.  However
we currently do that for all non-logical operations, regardless
whether they are replacement steps or not.

In particular a RefreshStep is non-logical, but it's also not a
replacement step.  We still want to show them during the diff because
their output can be important.  Especially if the user just requested
a diff it doesn't make sense to hide the diff from them at the same
time.

The intention here is to only hide replacement steps, so do that.