be consistent in the way we round seconds
In some cases we only display full seconds, particularly when
rendering the summary event, and when converting the engine event for
use in the API.  However we are currently not consistent with that,
rounding up in the display code and rounding down when converting the
engine event.

It probably doesn't matter what exactly we are doing here, but for
writing display snapshot tests it does.  Let's split the difference
here, and always round to the nearest second.