be consistent in the way we round seconds (#16070)
In some cases we only display full seconds, particularly when rendering
the summary event, and when converting the engine event for use in the
API. However we are currently not consistent with that, rounding up in
the display code and rounding down when converting the engine event.

It probably doesn't matter what exactly we are doing here, but for
writing display snapshot tests it does. Let's split the difference here,
and always round up, so we never give the impression that the operation
took no time.

It doesn't look like Pulumi Cloud shows this at all, so it probably
doesn't matter for users, and we could just round up for everyone as
well. I don't have strong opinions about that.