allow the regular CI pipeline to be run on external PRs
Currently when an external contributor submits a PR, we need to kick
of the CI pipeline manually.  Doing that runs basically the same tests
as the regular CI pipeline, except it's easy for it to get out of
sync.

The only advantage of doing it this way is that we can review the PR
before running the pipeline.  However the way our GitHub repo is set
up, we already need to approve PRs before CI is being run.

Unify the CI pipeline, which allows for the status to be reported on
the PR, making the experience for external contributors a bit nicer,
while still requiring us to approve running the CI jobs first but now
just with the click of a button.