implement programmatic default providers in the engine and Go SDK
Implement default providers in the engine and the Go SDK first (Go SDK
for no other reason other than I'm most familiar with that language).

The SDK sends a RegisterDefaultProvider request to the engine, which
adds the provider to its default provider map, and explicitly
disallows creating an implicit default provider going forward.

The locking around this happens on the SDK side, since we need to
disallow RegisterResource requests happening at the same time as
DefaultProvider requests.  Here we can simply take a read lock before
creating the RegisterResource goroutine, and a write lock for the
duration of the RegisterDefaultProvider call.

This way we can have RegisterResource calls happen in parallel, but
the RegisterDefaultProvider call will not go ahead before the read
lock is released.  Similarly, the next RegisterResource call will be
locked until the RegisterDefaultProvider call has finished.