Sort `requirements.txt` files when generating Python code (#16115)
A couple of our tests now generate multi-line `requirements.txt` whose
order is dependent on that in which we enumerate packages when testing.
This PR changes this so that we always sort them, meaning we can
consistently check against snapshotted results (as opposed to having
flaky tests that only pass when the generated order matches the one we
happened to snapshot last).

Fixes https://github.com/pulumi/pulumi/issues/16113