avoid showing repeated messages
Sometimes we show hundreds of repeated messages in the logs,
e.g. "waiting for quiescence; 3 outputs outstanding".  This just makes
the logs harder to read.  Since we need to wait for the test to finish
before any of this is shown anyway, only log each message once and
show a little blurb about how often the message has been
repeated.

(Note that for simplicity if the last message is repeated we don't
show the same blurb, but that should be okay since the last messages
we're logging are always different.)