truncate debug log output in conformance tests (#16117)
This debug log output can get way too long, and GitHub Actions logs
can't even display it anymore when there are test failures. Cut each
individual message short, with an escape hatch in case someone wants to
run with the full output locally.

I had a look at the test failure in
https://github.com/pulumi/pulumi/pull/16116 and it seems like it's still
having the same issue. Looking again at the test output it's actually
this t.Log that includes the registered resource outputs, that contains
the large amounts of data. I arbitrarily decided to make the cutoff 100
chars, but that could be adjusted.