make continue-on-error continue if preview fails
In the previous commit we made continue-on-error work with check
failures.  This means we now also ignore errors that can crop up at
preview time, however we will currently refuse to go ahead with the up
in such a case. This is confusing to the user, as they wanted to
continue despite errors.

Allow the up to continue even if there are errors at preview time if
the ContinueOnError flag is set.