Change Refresh to report Diff on desired state not outputs
Change refresh steps to compute a diff similar to what would happen if a `preview` were run immediately after the refresh using the inputs from the program, but inverted.

This doesn't change what is stored back into state, but does produce a diff that is more aligned with true "change to desired state".

This fixes several corner cases that are unfortuante in the prior implementation:
1. Changes to output-only proeprties - like `etag`s - are no longer shown as refresh diffs (but are updated in the state outputs).
2. IgnoreChanges now works to ignore property additions or changes (the changes are still updated in state so that future updates don't overwrite them again).

TODO:
- [ ] The diff display isn't updated yet - only the progres row display.
- [ ] Tests