Fix the testprovider to not always replace
The provider previously triggers replaces on every update, because it replaces on `prefix` but doesn't store it into outputs.  This *also* triggers the new refresh logic to report an update as part of the refresh, but it seems it is indeed just "broken" even outside of refresh, we just weren't testing multiple consecutive updates to these resources previously.