fix a panic when refresh is used with --continue-on-error
Refreshes have their own way of dealing with errors, which does not
match what we want to do with ContinueOnError (it's further reaching,
we still want to refresh dependents of failed resources).  Therefore
we made them mutually exclusive.

The original change however missed that it's also possible for
refreshes to happen during `pulumi up` using the `--refresh` flag.
This currently results in an assert/panic.

Fix this by not passing the `ContinueOnError` option through when we
have a refresh.  This still gives us the right behaviour since refresh
ignores the errors anyway.