fix a panic when refresh is used with --continue-on-error (#16184)
Refreshes have their own way of dealing with errors, which does not
match what we want to do with ContinueOnError (it's further reaching, we
still want to refresh dependents of failed resources). Therefore we made
them mutually exclusive.

The original change however missed that it's also possible for refreshes
to happen during `pulumi up` using the `--refresh` flag. This currently
results in an assert/panic.

Fix this by not passing the `ContinueOnError` option through when we
have a refresh. This still gives us the right behaviour since refresh
ignores the errors anyway.

Fixes https://github.com/pulumi/pulumi/issues/16177

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->