update dotnet version in tests to latest (#16198)
Update to the latest dotnet version. Pulumi.Awsx started depending on
that, so using an earlier version makes the tests fail.

This happened a few times now, so we should really fix this in a way
that doesn't require us to hardcode this version. But I want to get the
merge queue unblocked, and this is probably the quickest way to do that.