implement programmatic default providers based on context in the engine and Go
This PR takes an alternative approach for implementing programmatic
default providers, that was suggested in the internal design doc.

Instead of making the default provider global, we create an API with a
callback that creates a new subcontext.  In that subcontext and only
in that subcontext is the provider set as the default provider.  We
achieve this by setting up a new monitor with a different port, so the
engine can distinguish between the contexts.