Weaken Output.all types
The previous types stated that all the arguments to `Output.all` should
be of the same type, however you can pass through a list of mixed types.

If `Output.all` returned a tuple isntead, we could provide overloads
that returned a `Output[tuple[T1, T2, ...]]` with the precise types.
With the return type being a list, we are limited to `List[Any]`.