language_test: allow more characters in the log before truncating
We initially introduced this limit for log messages in
https://github.com/pulumi/pulumi/pull/16117.  However I've since
realized that these log messages also include useful
stacktraces (when the program being tested throws an exception), which
are a bit longer.

The intention here is really to avoid *really* long messages, so
increase the maximum message size a little to still avoid those, but
allow us to see more of the stacktrace that should hopefully be useful
for debugging.

The number is mostly still based on a guesstimate of what's useful,
but it's easy to tweak if we need to tweak it further.